requirements.txt
GPUtil
dicom2nifti
antspyx
-e git+https://github.com/MIC-DKFZ/HD-BET#egg=HD-BET

